Implemented ABTH for MSNP. The result is that an invite message is injected into the client TCP stream.
This implementation of the hack requires on average a little less than 2 seconds to execute,
and varies depending on the response time of packet travelling from the client to the server.
Not bad, but could use some improvement.

The logic of the hack remains the same as with the scripted implementation:
1. Wait for a ping response packet (QNG)
2. Build packets based on the TCP info of the QNG packet
3. First send pings to the server
4. Then send challenges to the client
5. Finally invite the client

Note for step 3 and 4, the response of each packet is the trigger for the next packet to be sent.
E.g. QNG must arrive for the next PNG to be sent. This ensures reliable packet injection,
although not the quickest. Otherwise, the server cannot process packets that fast.
Something may be worked on for the speed by estimating the response time of the server.

Updated some interfaces of PacketObservable as references make more sense than pointers.

MSNPABTH is pretty badly hacked up right now. It runs off the thread of PacketObserver,
which is fine for the current algorithm (1 packet sent per packet received).
It may need its own thread and a packet queueing mechanism in the future.